pyproject.toml
[project]
name = "eatellect-app"
version = "0.1.0"
description = "An AI-powered application for analyzing product health information from barcodes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyzbar>=0.1.9",
    "opencv-python>=4.8.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.300",
    "huggingface-hub>=0.17.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
namespaces = false

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
